Make base class of communicator
Nonblocking duplication of intracommunicator is also removed.
